+++ src/@ionic-native/plugins/barcode-scanner/index.ts
@@ -48,6 +48,11 @@ export interface BarcodeScannerOptions {
   torchOn?: boolean;
 
   /**
+   * Save scan history. Defaults to `false`. Supported on Android only.
+   */
+  saveHistory?: boolean;
+
+  /**
    * Display scanned text for X ms. 0 suppresses it entirely, default 1500. Supported on Android only.
    */
   resultDisplayDuration?: number;